Query Exist.io for all correlations for a specific attribute

Returns information on all correlations for a specific attribute.

exist_attribute_correlations(
  attribute = "mood",
  limit = NULL,
  page = 1,
  date_min = NULL,
  date_max = NULL,
  latest = NULL
)

Arguments

attribute
The attribute for which to get all correlations. Defaults to mood.

limit
Number of values to return per page, starting with today. Optional, max is 100.

page
Page index. Optional, default is 1.

date_min
Oldest date (inclusive) of results to be returned, in format YYYY-mm-dd. Optional.

date_max
Most recent date (inclusive) of results to be returned, in format YYYY-mm-dd. Optional.

latest
Set this to true to return only the most recently generated batch of correlations. Use this on its own without date_min and date_max.

Value

A list containing the information of all correlations for a specific attribute.

Details

This function expects the user to have the oauth token cached.

Examples

if (FALSE) {
exist_attribute_correlations()
}